% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{consolidateE}
\alias{consolidateE}
\title{Consolidate list of edgelists to a single edge list}
\usage{
consolidateE(listE)
}
\arguments{
\item{listE}{a list of edgelists}
}
\description{
Consolidate list of edgelists to a single edge list
}
