% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlin.R
\name{has}
\alias{has}
\title{What does the last edgelist or nodelist contain?}
\usage{
has(gremlin, key, query)
}
\arguments{
\item{gremlin}{a gremlin object}

\item{key}{string of attribute to query}

\item{query}{string that returns a boolean when evaluated}
}
\description{
What does the last edgelist or nodelist contain?
}
