% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlin.R
\name{makeGremlin}
\alias{makeGremlin}
\title{Create a Gremlin object}
\usage{
makeGremlin(graph, last = "vertices")
}
\arguments{
\item{graph}{graph object}

\item{last}{default query status of Gremlin}
}
\description{
Create a Gremlin object
}
