% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{queryE}
\alias{queryE}
\title{Ask the gremlin about it's edges}
\usage{
queryE(gremlin, mode)
}
\arguments{
\item{gremlin}{a gremlin object}

\item{mode}{'in', 'out', 'all', this is ignored for undirected graphs}
}
\description{
Find out which nodes connect with edges
}
