% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices.R
\name{queryV}
\alias{queryV}
\title{Ask the gremlin about it's nodes}
\usage{
queryV(gremlin, mode)
}
\arguments{
\item{gremlin}{a gremlin object}

\item{mode}{'in', 'out', 'all', 'total'}
}
\description{
Find out which nodes connect with which
}
